feat(shell): retire left ToolStripe rail; Inspector triggers move to toolbar
The unified sidebar navigator now owns primary navigation, so the redundant
left icon rail is gone. Git/Agents/Commands/Preview toggles relocate into the
titlebar's right cluster (same store actions, active states, shortcuts). One
clean navigation structure instead of rail + sidebar.
